+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v0.5.1] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [NPM v0.5.1] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.5.1 - previous-versions/sync-for-expenses-version-1
+
+## 2023-10-03 00:15:43
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.2] previous-versions/sync-for-expenses-version-1
+### Releases
+- [NPM v0.5.2] https://www.npmjs.com/package/@codat/sync-for-expenses-version-1/v/0.5.2 - previous-versions/sync-for-expenses-version-1